requirements.txt
# Core Pydantic AI and related packages
pydantic-ai[all]==0.0.14
pydantic>=2.10
pydantic-settings==2.4.0

# HTTP client and async support
httpx>=0.27.2
asyncio-throttle==1.0.2

# Database and data platforms
snowflake-connector-python==3.12.1
snowflake-sqlalchemy==1.6.1

# Microsoft Graph and Azure integration
azure-identity==1.19.0
msal==1.31.0
msgraph-core==1.2.0

# CLI and console interface
rich==13.8.0
click==8.1.7
typer==0.12.5

# Environment and configuration
python-dotenv==1.0.1

# Testing framework
pytest==8.3.3
pytest-asyncio==0.23.8
pytest-mock==3.14.0
pytest-cov==5.0.0

# Code quality and linting
ruff==0.6.5
mypy==1.11.2

# Logging and monitoring
structlog==24.4.0

# Data validation and serialization
email-validator==2.2.0

# Development and testing utilities
faker==30.0.0

# JSON and data manipulation
orjson==3.10.7

# Date and time utilities
python-dateutil==2.9.0